Move app into 'examples' directory.
