+++ examples/example-wasm/public/favicon.ico
