+++ examples/example-wasm/public/index.html
