+++ examples/example-wasm/public/logo192.png
