+++ examples/example-wasm/public/logo512.png
