+++ examples/example-wasm/public/manifest.json
