+++ examples/example-wasm/public/robots.txt
