+++ examples/example-wasm/src/AuthenticatedApp.tsx
