+++ examples/example-wasm/src/assets/logo.png
