+++ examples/example-wasm/src/assets/screenshot-realm-web-sync-tasks.png
