+++ examples/example-wasm/src/atlas-app-services/config.json
