+++ examples/example-wasm/src/components/IntroText.tsx
