+++ examples/example-wasm/src/components/NavBar.tsx
