+++ examples/example-wasm/src/components/PageLayout.tsx
