+++ examples/example-wasm/src/hooks/useAppManager.ts
