+++ examples/example-wasm/src/index.tsx
