+++ examples/example-wasm/src/pages/ErrorPage.tsx
