+++ examples/example-wasm/src/react-app-env.d.ts
