+++ examples/example-wasm/src/reportWebVitals.ts
