+++ examples/example-wasm/src/setupTests.ts
