+++ examples/example-wasm/src/styles/App.module.css
