+++ examples/example-wasm/src/styles/NavBar.module.css
