+++ examples/example-wasm/src/styles/global.css
