+++ package.json
@@ -47,7 +47,7 @@
   },
   "workspaces": [
     "example",
-    "example-wasm",
+    "examples/example-wasm",
     "packages/realm/bindgen/",
     "packages/realm/bindgen/vendor/realm-core/",
     "packages/babel-plugin",
